+++ babel.config.js
@@ -9,5 +9,8 @@ module.exports = {
       }
     ],
     '@babel/preset-typescript'
+  ],
+  plugins: [
+    '@babel/plugin-proposal-class-properties'
   ]
 };
