+++ packages/cli/src/index.ts
@@ -1,9 +1,8 @@
 import path from 'path';
 import { readdirSync, Dirent } from 'fs';
-import { getFolderSize, convertBytes } from './sizeUtils';
-import blacklisted from './blacklisted';
+import { convertBytes } from './sizeUtils';
 import { getAllNodeModules } from './finders';
-import { hasKey } from './utils';
+import { Problems } from './Problems';
 
 function getDirectories(source: string) {
   return readdirSync(source, { withFileTypes: true })
@@ -31,46 +30,6 @@ function isNamespaceDependency(source: string) {
 	return currentFolder.includes('@');
 }
 
-const getProblems = (name: string[]) => {
-	const subReport = {
-		problems: [],
-		totalSize: 0
-	};
-
-	name.forEach((dir: string) => {
-    const splittedFullPath = dir.split('/');
-    const dirName = splittedFullPath.pop();
-    const fullpathWithoutFilename = splittedFullPath.join('/');
-    if(dirName.includes('src')) {
-      const packageJson = require(`${fullpathWithoutFilename}/package.json`);
-	  const mainField = packageJson?.main;
-	  const devDeps = packageJson?.devDependencies;
-	  const hasFlow = hasKey(devDeps, 'flow-bin');
-
-	  // edge case for flow binaries, where the don't have any way
-	  // to get the type definitions as ts has
-      if(mainField?.includes('dist/') && !hasFlow){
-		subReport.problems = [...subReport.problems, "src"];
-		subReport.totalSize += getFolderSize(dirName);
-      }
-    }
-
-		const problemsFound = blacklisted
-			.filter((blackListed) => {
-				return blackListed === dirName;
-			})
-			.map((itemBlacklisted) => {
-				if (itemBlacklisted) {
-					const size = getFolderSize(dir);
-					subReport.totalSize += size;
-					return itemBlacklisted;
-				}
-			});
-		subReport.problems = [ ...subReport.problems, ...problemsFound ];
-	});
-	return subReport;
-};
-
 const cleanupDirName = (fullPath: string) => {
 	const splittedPath = fullPath.split('/');
 	const rootPackage = splittedPath[splittedPath.length - 3];
@@ -93,7 +52,7 @@ const mountGraph = (rootDir: string[]) => {
 	rootDir.forEach((dir: string) => {
     const subDirectories = getSubDirectories(dir);
 		if (!hasNMInside(dir)) {
-			const report = getProblems(subDirectories);
+			const { report } = new Problems(subDirectories);
 			if (report.problems.length) {
 				results.totalSaved += report.totalSize;
 				results.perPackage[cleanupDirName(dir)] = {
@@ -119,8 +78,8 @@ export function analyze(pathToNodeModules) {
 		perPackage: [],
 		totalSaved: 0
   };
-  
-	pathNM.forEach((nodePath) => {
+
+	pathNM.forEach((nodePath: string) => {
 		const initialDirs = getDirectories(nodePath);
 		const { perPackage, totalSaved } = mountGraph(initialDirs);
 		result.perPackage = { ...result.perPackage, ...perPackage };
